requirements.txt
# --- Core Framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
gunicorn>=21.2.0              # Production WSGI server for AWS
python-multipart>=0.0.12      # Critical for Image/Audio Uploads
python-dotenv>=1.0.1
httpx>=0.28.0                 # Async HTTP client

# --- AWS SDK ---
boto3>=1.34.0                 # S3, SNS, Lambda, IAM access

# --- Database (Vector Search) ---
qdrant-client>=1.13.0         # Qdrant Cloud + Local support

# --- Computer Vision ---
ultralytics>=8.3.0            # YOLOv11 / YOLOv26 Logic
transformers>=4.45.0          # Florence-2 Vision Agent
sentencepiece>=0.2.0          # Required by SigLIP tokenizer backend
timm>=1.0.9                   # Required for Florence-2 backbone
einops>=0.8.0                 # Required for Florence-2 tensor ops
torch>=2.4.0                  # PyTorch (Core AI Engine)
torchvision>=0.19.0           # Image processing for PyTorch
accelerate>=0.26.0            # For CUDA acceleration

# --- Embeddings & Image Processing ---
fastembed>=0.5.0              # Vector Embeddings (CPU + GPU)
pillow>=11.0.0                # Image manipulation
numpy>=1.26.0                 # Math ops

# --- Audio Intelligence ---
openai-whisper>=20231117      # Stable OpenAI Whisper release
soundfile>=0.13.0             # Audio file reading

# --- B2B Industrial Agent (PDF Parsing) ---
pymupdf>=1.25.0               # Reads PDF Manifests/Purchase Orders
pypdf>=5.0.0                  # PDF Reader (fallback)

# --- Advanced Reasoning ---
google-genai>=1.0.0           # Gemini API